% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{metaScreen}
\alias{metaScreen}
\title{metRscreen}
\usage{
metaScreen()
}
\value{
A dataframe of decisioned papers
}
\description{
The metRscreen shiny app allows you to screen papers via their abstracts and titles and allows for highlighting of keywords in multiple colours.
}
