% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{metRscreen}
\alias{metRscreen}
\title{Run the metRscreen paper screening app}
\usage{
metRscreen(
  screen.file,
  reject.list = NULL,
  collab.names = NULL,
  keywords = list(green = NULL, red = NULL, purple = NULL, orange = NULL, blue = NULL)
)
}
\arguments{
\item{screen.file}{path to the csv file containing references you wish to screen.}

\item{reject.list}{list of rejection reasons to be added to metRscreen, can be left empty}

\item{collab.names}{vector of screener names to switch on collaborative mode, can be left empty.
Each screener gets their own files (\verb{<screen.file>_<name>_Screened.csv} and
\verb{<screen.file>_<name>_history.rds}), you switch screener in the app, and other
screeners' decisions can be shown or hidden. A combined
\verb{<screen.file>_Collab_Summary.csv} flags agreements and conflicts. Names are
remembered, so later sessions can leave this empty or add new screeners.}

\item{keywords}{takes a list of green, red, purple, orange, or blue keywords to add to the highlight word command.}
}
\value{
A dataframe of decisioned papers
}
\description{
The metRscreen shiny app allows you to screen papers via their abstracts and titles and allows for highlighting of keywords in multiple colours.
}
