% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{metRscreen}
\alias{metRscreen}
\title{metRscreen}
\usage{
metRscreen(reject.list = NULL, metRDS = NULL)
}
\arguments{
\item{reject.list}{list of rejection reasons to be added to metRscreen, can be left empty}

\item{metRDS}{a save state file path to resume screening from a previous metRscreen screen}
}
\value{
A dataframe of decisioned papers
}
\description{
The metRscreen shiny app allows you to screen papers via their abstracts and titles and allows for highlighting of keywords in multiple colours.
}
