% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{metRscreen}
\alias{metRscreen}
\title{metRscreen}
\usage{
metRscreen(screen.file, reject.list = NULL)
}
\arguments{
\item{screen.file}{path to the csv file containing references you wish to screen.}

\item{reject.list}{list of rejection reasons to be added to metRscreen, can be left empty}
}
\value{
A dataframe of decisioned papers
}
\description{
The metRscreen shiny app allows you to screen papers via their abstracts and titles and allows for highlighting of keywords in multiple colours.
}
